+++ Projects/CoX/Common/Messages/Map/StoresEvents.h
@@ -5,7 +5,7 @@
  * This software is licensed under the terms of the 3-clause BSD License. See LICENSE.md for details.
  */
 
-
+#pragma once
 
 #include "GameCommand.h"
 #include "MapEventTypes.h"
