+++ Projects/CoX/Servers/MapServer/ScriptingEngine/CMakeLists.txt
@@ -1,22 +1,15 @@
-if(ENABLE_SCRIPTING_ENGINE)
-    set(SCRIPTING_CPP ScriptingEngine.cpp)
-    set(SCRIPTING_REGISTERTYPES_CPP ScriptingEngine_RegisterTypes.cpp)
-    set(SCRIPTING_HELPERS_CPP ScriptingEngine_Helpers.cpp)
-    set(SCRIPTING_PRIVATE_CPP ScriptingEnginePrivate.cpp)
-    set(SCRIPTING_PRIVATE_H ScriptingEnginePrivate.h)
-else()
-    set(SCRIPTING_CPP ScriptingEngine_Null.cpp)
-endif()
 
 SET(target_CPP
-    ${SCRIPTING_CPP}
-    ${SCRIPTING_REGISTERTYPES_CPP}
-    ${SCRIPTING_HELPERS_CPP}
-    ${SCRIPTING_PRIVATE_CPP}
+    ScriptingEngine.cpp
+    ScriptingEngine_CharacterTypes.cpp
+    ScriptingEngine_GenericTypes.cpp
+    ScriptingEngine_Helpers.cpp
+    ScriptingEngine_SpawnerTypes.cpp
+    ScriptingEnginePrivate.cpp
 )
 SET(target_INCLUDE
     ScriptingEngine.h
-    ${SCRIPTING_PRIVATE_H}
+    ScriptingEnginePrivate.h
 )
 
 SET(target_INCLUDE_DIR ${CMAKE_CURRENT_SOURCE_DIR})
@@ -31,10 +24,8 @@ ${target_INCLUDE}
 add_library(ScriptingEngine ${target_SOURCES})
 target_link_libraries(ScriptingEngine PUBLIC Qt5::Core gameData)
 
-if(ENABLE_SCRIPTING_ENGINE)
-    target_compile_definitions(ScriptingEngine PUBLIC -DSCRIPTING_ENABLED=1)
-    target_link_libraries(ScriptingEngine INTERFACE lua_IMP sol2_IMP)
-endif()
+target_compile_definitions(ScriptingEngine PUBLIC -DSCRIPTING_ENABLED=1)
+target_link_libraries(ScriptingEngine INTERFACE lua_IMP sol2_IMP)
 
 if(ENABLE_TESTS)
 #    add_subdirectory(UnitTests)
