+++ Projects/CoX/Servers/MapServer/ScriptingEngine/ScriptingEngine.cpp
@@ -6,16 +6,7 @@
  */
 
 #include "ScriptingEngine.h"
-#include "MapInstance.h"
-#include "GameData/playerdata_definitions.h"
-#include "DataHelpers.h"
 
-#include <QtCore/QFileInfo> // for include support
-#include <QtCore/QDir>
-#include <QtCore/QDebug>
-
-using namespace SEGSEvents;
-//static constexpr const int MAX_INCLUDED_FILE_SIZE=1024*1024; // 1MB of lua code should be enough for anyone :P
 int luaopen_package(lua_State *)
 {
     assert(false && "functionality removed");
@@ -41,15 +32,11 @@ ScriptingEngine::ScriptingEngine()
 
 ScriptingEngine::~ScriptingEngine() = default;
 
-
 void ScriptingEngine::registerTypes()
 {
     register_GenericTypes();
     register_SpawnerTypes();
     register_CharacterTypes();
-
 }
-
-
 
 //! @}
