+++ Projects/CoX/Servers/MapServer/ScriptingEngine/ScriptingEnginePrivate.h
@@ -5,6 +5,8 @@
  * This software is licensed under the terms of the 3-clause BSD License. See LICENSE.md for details.
  */
 
+#pragma once
+
 #include <QString>
 #include <QSet>
 #include <sol2/sol.hpp>
